requirements.txt
google-generativeai
python-dotenv
streamlit
plotly


streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.38
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0
python-dotenv==1.0.1
xgboost==2.0.3
ta==0.11.0